refactor: splitted the code in three files, two utils file and a index module file
